+++ rabbithole-derive/src/error.rs
@@ -5,6 +5,11 @@ pub enum EntityDecoratorError {
          `#[entity(type = \"foo_type\")]`"
     )]
     InvalidEntityType,
+    #[error(
+        "`EntityDecorator` needs a Data Service to handle the CRUD operations, using \
+         `#[entity(service(FooService))]` to register one"
+    )]
+    LackOfService,
     #[error("Duplicated Id fields detected")]
     DuplicatedId,
     #[error("Invalid unit decorator {0}, the valid ones: [id, to_one, to_many]")]