+++ rabbithole-endpoint-actix/src/lib.rs
@@ -3,21 +3,25 @@ pub mod settings;
 use actix_web::http::{header, HeaderMap, StatusCode};
 use actix_web::web;
 use actix_web::{HttpRequest, HttpResponse};
-use futures::{FutureExt, TryFutureExt};
-use rabbithole::entity::SingleEntity;
+use rabbithole::entity::{Entity, SingleEntity};
 
 use crate::settings::{ActixSettingsModel, JsonApiSettings};
 use actix_web::dev::HttpResponseBuilder;
 
 use rabbithole::model::error;
 use rabbithole::model::version::JsonApiVersion;
-use rabbithole::operation::Fetching;
+use rabbithole::operation::{
+    Creating, Deleting, Fetching, IdentifierDataWrapper, OperationResultData, ResourceDataWrapper,
+};
 use rabbithole::rule::RuleDispatcher;
 use rabbithole::JSON_API_HEADER;
 use serde::export::TryFrom;
 
-use rabbithole::query::Query;
-use std::marker::PhantomData;
+use futures::lock::Mutex;
+use rabbithole::query::QuerySettings;
+
+use rabbithole::model::document::Document;
+use std::sync::Arc;
 
 fn error_to_response(err: error::Error) -> HttpResponse {
     new_json_api_resp(
@@ -27,157 +31,280 @@ fn error_to_response(err: error::Error) -> HttpResponse {
 }
 
 #[derive(Debug, Clone)]
-pub struct ActixSettings<T>
-where
-    T: 'static + Fetching,
-{
+pub struct ActixSettings {
     pub path: String,
     pub uri: url::Url,
     pub jsonapi: JsonApiSettings,
-    _data: PhantomData<T>,
+    pub query: QuerySettings,
 }
 
-impl<T> TryFrom<ActixSettingsModel> for ActixSettings<T>
-where
-    T: 'static + Fetching + Send + Sync,
-    T::Item: Send + Sync,
-{
+impl TryFrom<ActixSettingsModel> for ActixSettings {
     type Error = url::ParseError;
 
     fn try_from(value: ActixSettingsModel) -> Result<Self, Self::Error> {
-        let ActixSettingsModel { host, port, path, jsonapi } = value;
+        let ActixSettingsModel { host, port, path, jsonapi, query } = value;
         let uri = format!("http://{}:{}", host, port).parse::<url::Url>().unwrap();
         let uri = uri.join(&path).unwrap();
-        Ok(Self { path, uri, jsonapi, _data: PhantomData })
+        Ok(Self { path, uri, jsonapi, query })
     }
 }
 
-impl<T> ActixSettings<T>
-where
-    T: 'static + Fetching + Send + Sync,
-    T::Item: Send + Sync,
-{
-    pub fn fetch_collection(
-        self, req: HttpRequest,
-    ) -> impl futures01::Future<Item = HttpResponse, Error = actix_web::Error> {
-        if let Err(err_resp) = check_header(&self.jsonapi.version, &req.headers()) {
-            return futures::future::ok(err_resp).boxed_local().compat();
+macro_rules! to_response {
+    (Resource: $this:ident, $item:ident) => {{
+        let rabbithole::operation::OperationResultData { data, additional_links, additional_meta } =
+            $item;
+        let resource = data.to_resource(&$this.uri.to_string(), &Default::default());
+        match resource {
+            Some(mut resource) => {
+                resource.extend_meta(additional_meta);
+                resource.extend_links(additional_links);
+                Ok(actix_web::HttpResponse::Ok()
+                    .json(rabbithole::operation::ResourceDataWrapper { data: resource }))
+            },
+            None => Ok(actix_web::HttpResponse::NoContent().finish()),
         }
-        match Query::from_uri(req.uri()) {
-            Ok(query) => {
-                let fut = async move {
-                    let vec_res = T::fetch_collection(&query).await;
-                    match vec_res {
-                        Ok(vec) => {
-                            match T::vec_to_document(
-                                &vec,
-                                &self.uri.to_string(),
-                                &query,
-                                &req.uri().into(),
-                            )
-                            .await
-                            {
-                                Ok(doc) => Ok(HttpResponse::Ok().json(doc)),
-                                Err(err) => Ok(error_to_response(err)),
-                            }
-                        },
+    }};
+    (Relationship: $this:ident, $item:ident) => {{
+        let rabbithole::operation::OperationResultData { data, additional_links, additional_meta } =
+            $item;
+        let (field_name, item) = data;
+        let resource = item.to_resource(&$this.uri.to_string(), &Default::default());
+        match resource {
+            Some(mut resource) => {
+                resource.extend_meta(additional_meta);
+                resource.extend_links(additional_links);
+
+                Ok(actix_web::HttpResponse::Ok().json(
+                    rabbithole::operation::IdentifierDataWrapper {
+                        data: resource.relationships.get(&field_name).cloned().unwrap().data,
+                    },
+                ))
+            },
+            None => Ok(actix_web::HttpResponse::NoContent().finish()),
+        }
+    }};
+}
+
+macro_rules! single_step_operation {
+    ($return_ty:ident:  $fn_name:ident, $mark:ident, $( $param:ident => $ty:ty ),+) => {
+        pub async fn $fn_name<T>(this: web::Data<Self>, service: actix_web::web::Data<std::sync::Arc<futures::lock::Mutex<T>>>, req: actix_web::HttpRequest, $($param: $ty),+)
+          -> Result<actix_web::HttpResponse, actix_web::Error>
+            where
+                T: 'static + rabbithole::operation::$mark + Send + Sync,
+                T::Item: rabbithole::entity::SingleEntity + Send + Sync,
+          {
+            if let Err(err_resp) = check_header(&this.jsonapi.version, &req.headers()) {
+                return Ok(err_resp);
+            }
+
+            match service.lock().await.$fn_name($(&$param.into_inner()),+).await {
+                Ok(item) => {
+                    to_response!($return_ty: this, item)
+                },
+                Err(err) => Ok(error_to_response(err)),
+            }
+        }
+    };
+}
+
+impl ActixSettings {
+    single_step_operation!(Resource: update_resource, Updating, params => web::Path<String>, body => web::Json<ResourceDataWrapper>);
+
+    single_step_operation!(Relationship: replace_relationship, Updating, params => web::Path<(String, String)>, body => web::Json<IdentifierDataWrapper>);
+
+    single_step_operation!(Relationship: add_relationship, Updating, params => web::Path<(String, String)>, body => web::Json<IdentifierDataWrapper>);
+
+    single_step_operation!(Relationship: remove_relationship, Updating, params => web::Path<(String, String)>, body => web::Json<IdentifierDataWrapper>);
+}
+
+impl ActixSettings {
+    pub async fn delete_resource<T>(
+        this: web::Data<Self>, service: web::Data<Arc<Mutex<T>>>, params: web::Path<String>,
+        req: actix_web::HttpRequest,
+    ) -> Result<HttpResponse, actix_web::Error>
+    where
+        T: 'static + Deleting + Send + Sync,
+        T::Item: rabbithole::entity::SingleEntity + Send + Sync,
+    {
+        if let Err(err_resp) = check_header(&this.jsonapi.version, &req.headers()) {
+            return Ok(err_resp);
+        }
+
+        match service.lock().await.delete_resource(&params.into_inner()).await {
+            Ok(OperationResultData { additional_links, additional_meta, .. }) => {
+                if additional_links.is_empty() && additional_meta.is_empty() {
+                    Ok(actix_web::HttpResponse::NoContent().finish())
+                } else {
+                    Ok(actix_web::HttpResponse::Ok()
+                        .json(Document::null(additional_links, additional_meta)))
+                }
+            },
+            Err(err) => Ok(error_to_response(err)),
+        }
+    }
+}
+
+impl ActixSettings {
+    pub async fn create<T>(
+        this: web::Data<Self>, service: web::Data<Arc<Mutex<T>>>, req: actix_web::HttpRequest,
+        body: web::Json<ResourceDataWrapper>,
+    ) -> Result<HttpResponse, actix_web::Error>
+    where
+        T: 'static + Creating + Send + Sync,
+        T::Item: rabbithole::entity::SingleEntity + Send + Sync,
+    {
+        if let Err(err_resp) = check_header(&this.jsonapi.version, &req.headers()) {
+            return Ok(err_resp);
+        }
+
+        match service.lock().await.create(&body.into_inner()).await {
+            Ok(OperationResultData { data, additional_links, additional_meta }) => {
+                let mut resource =
+                    data.to_resource(&this.uri.to_string(), &Default::default()).unwrap();
+                resource.extend_meta(additional_meta);
+                resource.extend_links(additional_links);
+                Ok(actix_web::HttpResponse::Ok().json(ResourceDataWrapper { data: resource }))
+            },
+            Err(err) => Ok(error_to_response(err)),
+        }
+    }
+}
+
+impl ActixSettings {
+    pub async fn fetch_collection<T>(
+        this: web::Data<Self>, service: web::Data<Arc<Mutex<T>>>, req: HttpRequest,
+    ) -> Result<HttpResponse, actix_web::Error>
+    where
+        T: 'static + Fetching + Send + Sync,
+        T::Item: rabbithole::entity::SingleEntity + Send + Sync,
+    {
+        if let Err(err_resp) = check_header(&this.jsonapi.version, &req.headers()) {
+            return Ok(err_resp);
+        }
+        match this.query.from_uri(req.uri()) {
+            Ok(query) => match service.lock().await.fetch_collection(&query).await {
+                Ok(OperationResultData { data, additional_links, additional_meta }) => {
+                    match data.to_document(
+                        &this.uri.to_string(),
+                        &query,
+                        req.uri().into(),
+                        additional_links,
+                        additional_meta,
+                    ) {
+                        Ok(doc) => Ok(HttpResponse::Ok().json(doc)),
                         Err(err) => Ok(error_to_response(err)),
                     }
-                };
-
-                fut.boxed_local().compat()
+                },
+                Err(err) => Ok(error_to_response(err)),
             },
-            Err(err) => futures::future::ok(error_to_response(err)).boxed_local().compat(),
+            Err(err) => Ok(error_to_response(err)),
         }
     }
 
-    pub fn fetch_single(
-        self, param: web::Path<String>, req: HttpRequest,
-    ) -> impl futures01::Future<Item = HttpResponse, Error = actix_web::Error> {
-        if let Err(err_resp) = check_header(&self.jsonapi.version, &req.headers()) {
-            return futures::future::ok(err_resp).boxed_local().compat();
+    pub async fn fetch_single<T>(
+        this: web::Data<Self>, service: web::Data<Arc<Mutex<T>>>, param: web::Path<String>,
+        req: HttpRequest,
+    ) -> Result<HttpResponse, actix_web::Error>
+    where
+        T: 'static + Fetching + Send + Sync,
+        T::Item: rabbithole::entity::SingleEntity + Send + Sync,
+    {
+        if let Err(err_resp) = check_header(&this.jsonapi.version, &req.headers()) {
+            return Ok(err_resp);
         }
-        match Query::from_uri(req.uri()) {
+        match this.query.from_uri(req.uri()) {
             Ok(query) => {
-                let fut = async move {
-                    match T::fetch_single(&param.into_inner(), &query).await {
-                        Ok(item) => {
-                            match item.to_document_automatically(
-                                &self.uri.to_string(),
-                                &query,
-                                &req.uri().into(),
-                            ) {
-                                Ok(doc) => Ok(new_json_api_resp(StatusCode::OK).json(doc)),
-                                Err(err) => Ok(error_to_response(err)),
-                            }
-                        },
-                        Err(err) => Ok(error_to_response(err)),
-                    }
-                };
-
-                fut.boxed_local().compat()
+                match service.lock().await.fetch_single(&param.into_inner(), &query).await {
+                    Ok(OperationResultData { data, additional_links, additional_meta }) => {
+                        match SingleEntity::to_document(
+                            &data,
+                            &this.uri.to_string(),
+                            &query,
+                            req.uri().into(),
+                            additional_links,
+                            additional_meta,
+                        ) {
+                            Ok(doc) => Ok(new_json_api_resp(StatusCode::OK).json(doc)),
+                            Err(err) => Ok(error_to_response(err)),
+                        }
+                    },
+                    Err(err) => Ok(error_to_response(err)),
+                }
             },
-            Err(err) => futures::future::ok(error_to_response(err)).boxed_local().compat(),
+            Err(err) => Ok(error_to_response(err)),
         }
     }
 
-    pub fn fetch_relationship(
-        self, param: web::Path<(String, String)>, req: HttpRequest,
-    ) -> impl futures01::Future<Item = HttpResponse, Error = actix_web::Error> {
-        if let Err(err_resp) = check_header(&self.jsonapi.version, &req.headers()) {
-            return futures::future::ok(err_resp).boxed_local().compat();
+    pub async fn fetch_relationship<T>(
+        this: web::Data<Self>, service: web::Data<Arc<Mutex<T>>>,
+        param: web::Path<(String, String)>, req: HttpRequest,
+    ) -> Result<HttpResponse, actix_web::Error>
+    where
+        T: 'static + Fetching + Send + Sync,
+        T::Item: rabbithole::entity::SingleEntity + Send + Sync,
+    {
+        if let Err(err_resp) = check_header(&this.jsonapi.version, &req.headers()) {
+            return Ok(err_resp);
         }
-        match Query::from_uri(req.uri()) {
+        match this.query.from_uri(req.uri()) {
             Ok(query) => {
                 let (id, related_field) = param.into_inner();
-                let fut = async move {
-                    match T::fetch_relationship(
+                match service
+                    .lock()
+                    .await
+                    .fetch_relationship(
                         &id,
                         &related_field,
-                        &self.uri.to_string(),
+                        &this.uri.to_string(),
                         &query,
                         &req.uri().into(),
                     )
                     .await
-                    {
-                        Ok(item) => Ok(new_json_api_resp(StatusCode::OK).json(item)),
-                        Err(err) => Ok(error_to_response(err)),
-                    }
-                };
-
-                fut.boxed_local().compat()
+                {
+                    Ok(OperationResultData { mut data, additional_links, additional_meta }) => {
+                        data.extend_links(additional_links);
+                        data.extend_meta(additional_meta);
+                        Ok(new_json_api_resp(StatusCode::OK).json(data))
+                    },
+                    Err(err) => Ok(error_to_response(err)),
+                }
             },
-            Err(err) => futures::future::ok(error_to_response(err)).boxed_local().compat(),
+            Err(err) => Ok(error_to_response(err)),
         }
     }
 
-    pub fn fetch_related(
-        self, param: web::Path<(String, String)>, req: HttpRequest,
-    ) -> impl futures01::Future<Item = HttpResponse, Error = actix_web::Error> {
-        if let Err(err_resp) = check_header(&self.jsonapi.version, &req.headers()) {
-            return futures::future::ok(err_resp).boxed_local().compat();
+    pub async fn fetch_related<T>(
+        this: web::Data<Self>, service: web::Data<Arc<Mutex<T>>>,
+        param: web::Path<(String, String)>, req: HttpRequest,
+    ) -> Result<HttpResponse, actix_web::Error>
+    where
+        T: 'static + Fetching + Send + Sync,
+        T::Item: rabbithole::entity::SingleEntity + Send + Sync,
+    {
+        if let Err(err_resp) = check_header(&this.jsonapi.version, &req.headers()) {
+            return Ok(err_resp);
         }
 
-        match Query::from_uri(req.uri()) {
+        match this.query.from_uri(req.uri()) {
             Ok(query) => {
                 let (id, related_field) = param.into_inner();
-                let fut = async move {
-                    match T::fetch_related(
+                match service
+                    .lock()
+                    .await
+                    .fetch_related(
                         &id,
                         &related_field,
-                        &self.uri.to_string(),
+                        &this.uri.to_string(),
                         &query,
                         &req.uri().into(),
                     )
                     .await
-                    {
-                        Ok(item) => Ok(new_json_api_resp(StatusCode::OK).json(item)),
-                        Err(err) => Ok(error_to_response(err)),
-                    }
-                };
-                fut.boxed_local().compat()
+                {
+                    Ok(item) => Ok(new_json_api_resp(StatusCode::OK).json(item)),
+                    Err(err) => Ok(error_to_response(err)),
+                }
             },
-            Err(err) => futures::future::ok(error_to_response(err)).boxed_local().compat(),
+            Err(err) => Ok(error_to_response(err)),
         }
     }
 }