+++ rabbithole-endpoint-actix/src/settings.rs
@@ -1,4 +1,5 @@
 use rabbithole::model::version::JsonApiVersion;
+use rabbithole::query::QuerySettings;
 use serde::Deserialize;
 
 #[derive(Debug, Deserialize, Clone)]
@@ -7,6 +8,7 @@ pub struct ActixSettingsModel {
     pub port: u32,
     pub path: String,
     pub jsonapi: JsonApiSettings,
+    pub query: QuerySettings,
 }
 
 #[derive(Debug, Deserialize, Clone)]