+++ rabbithole-endpoint-actix/tests/common/mod.rs
@@ -1,2 +1,2 @@
 /// https://jsonapi.org/format/#fetching-resources
-pub mod fetching_test;
+pub mod integration_test;