+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## [Unreleased]
 
+### Changed
+
+- Update `actix-web` to `2.0` with the async handler
+- A in-memory demo
+- `204 No Content`/`200 OK` when `PATCH`, `POST` and `DELETE` the resource
+- From `to_document_automatically` to `to_document`, now this function only handle a simple job - convert `Entity` to `Document`, no more, no less
+- Now User can return the additional links and meta info in `Operation` trait
+
 ## [0.3.0] - 2019-11-17
 
 ### Fixed